+++ CHANGELOG.rst
@@ -42,8 +42,7 @@ Added
 - Instance will not produce suspects during ``RecoveringSnapshot`` and
   ``BootstrappingBox``.
 
-- ``rst/BuildImages.cmake`` to generate images based on uml diagrams
-  from ``rst/uml``.
+- New script ``rst/BuildUML.cmake`` for rendering UML diagrams from the doc.
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Changed